requirements.txt.txt
django==3.2
pdfminer.six==20231228
langchain==0.2.14
reportlab==4.2.2